Refactor graphML: move to a child dir
